Move FedoraData to models.
